@@ -12,55 +12,55 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "^6.0.0",
-    "@angular/common": "^6.0.0",
-    "@angular/compiler": "^6.0.0",
-    "@angular/core": "^6.0.0",
-    "@angular/forms": "^6.0.0",
+    "@angular/animations": "^8.0.0",
+    "@angular/common": "^8.0.0",
+    "@angular/compiler": "^8.0.0",
+    "@angular/core": "^8.0.0",
+    "@angular/forms": "^8.0.0",
     "@angular/http": "^6.0.0",
-    "@angular/platform-browser": "^6.0.0",
-    "@angular/platform-browser-dynamic": "^6.0.0",
-    "@angular/router": "^6.0.0",
+    "@angular/platform-browser": "^8.0.0",
+    "@angular/platform-browser-dynamic": "^8.0.0",
+    "@angular/router": "^8.0.0",
     "@types/jasmine": "^2.8.7",
     "@types/jasminewd2": "^2.0.3",
     "@types/jest": "^22.2.3",
     "@types/lodash": "^4.14.108",
     "angular-font-awesome": "^3.1.2",
     "angular2-fontawesome": "^5.2.1",
-    "bootstrap": "^4.1.1",
+    "bootstrap": "^4.3.1",
     "core-js": "^2.5.4",
     "font-awesome": "^4.7.0",
     "intl": "^1.2.5",
     "istanbul-reports": "^1.3.0",
     "karma-coverage-istanbul-reporter": "^2.0.0",
-    "lodash": "^4.17.10",
+    "lodash": "^4.17.11",
+    "moment": "^2.24.0",
     "ng-autosize": "^1.1.0",
-    "ng2-materialize": "^1.8.0",
-    "ngx-bootstrap": "^2.0.5",
-    "ngx-moment": "^2.0.0",
+    "ng2-materialize": "^1.8.1",
+    "ngx-bootstrap": "^4.2.0",
+    "ngx-moment": "^3.4.0",
     "ngx-webstorage": "^2.0.1",
     "node-forge": "^0.7.5",
-    "rxjs": "^6.0.0",
-    "rxjs-compat": "^6.1.0",
-    "zone.js": "^0.8.26"
+    "rxjs": "^6.5.2",
+    "zone.js": "^0.9.1"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "~0.6.1",
-    "@angular/cli": "~6.0.1",
-    "@angular/compiler-cli": "^6.0.0",
-    "@angular/language-service": "^6.0.0",
+    "@angular-devkit/build-angular": "^0.800.2",
+    "@angular/cli": "~8.0.2",
+    "@angular/compiler-cli": "^8.0.0",
+    "@angular/language-service": "^8.0.0",
     "@types/node": "~8.9.4",
-    "codelyzer": "~4.2.1",
+    "codelyzer": "~5.1.0",
     "jasmine-core": "^3.1.0",
-    "karma": "^2.0.2",
+    "karma": "^4.1.0",
     "karma-cli": "^1.0.1",
     "karma-jasmine": "^1.1.2",
     "karma-jasmine-html-reporter": "^1.1.0",
     "karma-phantomjs-launcher": "^1.0.4",
     "phantomjs-prebuilt": "^2.1.16",
-    "protractor": "^5.3.2",
+    "protractor": "^6.0.0",
     "ts-node": "~5.0.1",
     "tslint": "~5.9.1",
-    "typescript": "~2.7.2"
+    "typescript": "^3.4.5"
   }
 }
